refactor(test): auth 라우트 테스트 리팩터 #313
