+++ src/testRoutes/testPages/thepott/apiTest/ThePottApiTestPage.jsx
@@ -10,6 +10,7 @@ import { FullScreen, Vstack } from "../../../../package/layout";
 import FlexOneContainer from "../../../../package/flexOneContainer/FlexOneContainer";
 import CustomButton from "../../../../package/customButton/CustomButton";
 import RoundBox from "../../../../package/RoundBox";
+import testAuth from "./testAuth";
 
 const RoundBoxGlobalShadow = ({ style, children, ...props }) => {
     return (
@@ -22,28 +23,6 @@ const RoundBoxGlobalShadow = ({ style, children, ...props }) => {
 const baseURL = import.meta.env.VITE_BASE_URL;
 const getHome = () => getThenLog(`${baseURL}`);
 const getHealth = () => getThenLog(`${baseURL}/health`);
-const postEmailVerification = () =>
-    postThenLog(`${baseURL}/api/auth/send-verification-email`, {
-        email: "nusilite@gmail.com",
-    });
-const postFanLogin = (callback) =>
-    postThenLog(
-        `${baseURL}/api/auth/login`,
-        {
-            email: "fan_dummy_1@gmail.com",
-            password: "fan123!",
-        },
-        callback,
-    );
-const postCompanyLogin = (callback) =>
-    postThenLog(
-        `${baseURL}/api/auth/login`,
-        {
-            email: "sm_dummy@company.com",
-            password: "company123!",
-        },
-        callback,
-    );
 
 const postFanPost = () =>
     axiosReturnsData("POST", "/api/posts", {
@@ -53,9 +32,9 @@ const postFanPost = () =>
 
 const ThePottApiTestPage = () => {
     const [accessToken, setAccessToken] = useState(null);
-    const [user, setUser] = useState(null);
 
-    const getMe = (callback) => getThenLog(`${baseURL}/api/auth/me`, callback, accessToken);
+    const testAuthReturns = testAuth(baseURL, accessToken, setAccessToken);
+
     const getFanSubscriptionWithName = (callback) =>
         getThenLog(`${baseURL}/api/subscriptions/?include_image=true`, callback, accessToken);
     const getFanSubscription = (callback) =>
@@ -93,17 +72,6 @@ const ThePottApiTestPage = () => {
             accessToken,
         );
 
-    const callbackLogin = (data) => {
-        setAccessToken(data.access_token);
-    };
-
-    const callbackMe = (data) => {
-        setUser(data);
-    };
-    const callbackLog = (data) => {
-        console.log({ data });
-    };
-
     const handleBulkUpload = (event) => {
         event.preventDefault();
         const formData = new FormData(event.target);
@@ -120,20 +88,17 @@ const ThePottApiTestPage = () => {
                         </RoundBoxGlobalShadow>
                         <CustomButton onClick={getHome}>get home</CustomButton>
                         <CustomButton onClick={getHealth}>get health</CustomButton>
-                        <CustomButton onClick={postEmailVerification}>
+                        <CustomButton onClick={testAuthReturns.postEmailVerification}>
                             send email verification
                         </CustomButton>
-                        <CustomButton onClick={() => postFanLogin(callbackLogin)}>
+                        <CustomButton onClick={testAuthReturns.postFanLogin}>
                             <p>fan login</p>
-                            <p>{accessToken}</p>
                         </CustomButton>
-                        <CustomButton onClick={() => postCompanyLogin(callbackLogin)}>
+                        <CustomButton onClick={testAuthReturns.postCompanyLogin}>
                             <p>company login</p>
-                            <p>{accessToken}</p>
                         </CustomButton>
-                        <CustomButton onClick={() => getMe(callbackMe)}>
+                        <CustomButton onClick={testAuthReturns.getMe}>
                             <p>get me</p>
-                            <p>{JSON.stringify(user)}</p>
                         </CustomButton>
                         <CustomButton onClick={() => getEvents(callbackLog)}>
                             <p>get events</p>
